refactor: add answer, content, reply props to FillTheGap block
