+++ front/src/pages/User/LearnPage/BlockElement/FillTheGap/GapsInput/GapsInput.styled.js
