+++ front/src/pages/User/LearnPage/BlockElement/FillTheGap/GapsInput/index.js
@@ -1,0 +1,3 @@
+import GapsInput from './GapsInput';
+
+export default GapsInput;
