--- front/src/pages/User/LearnPage/BlockElement/FillTheGap/Inputs/index.js
@@ -1,3 +1,0 @@
-import Inputs from './Inputs';
-
-export default Inputs;
